requirements.txt
-i https://pypi.org/simple
attrs==19.3.0
cycler==0.10.0
dataclasses==0.6
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
patsy==0.5.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
statsmodels==0.10.2
